[#385] Introduce & implement NetworkMagic
NetworkMagic allows us to differentiate between Addresses which
come from different networks. This prevents users from accidentally
sending funds from a mainnet wallet to a testnet address, which would
technically be a valid address on mainnet, but would be controlled by
unknown keys not under the user's control (thus rendering the funds
unusable). When addresses are created, the @ProtocolMagicId@ of the
machine is inserted into the @NetworkMagic@, which is added to the
@Address@ datatype.

In this commit I introduce NetworkMagic, wire it up to @ProtocolMagic@
in the configuration, and modify functions so that @NetworkMagic@ can
be passed down to both address creation and transaction validation. I
add a new constructor to @TxValidationError@
- `TxValidationNetworkMagicMismatch` - which is thrown when we
encounter a @NetworkMagic@ coming from an address which doesn't match
our cluster's @NetworkMagic@.